agents/documentation_agent.py
from dataclasses import dataclass
from datetime import datetime
from pathlib import Path
from typing import Dict, Optional, Any, List
from uuid import UUID

from pydantic import BaseModel, Field

from .base import BaseAgent, AgentType, AgentConfig, AgentContext

class DocumentationAgentConfig(AgentConfig):
    """Configuration specific to the Documentation Agent"""
    template_dir: Path = Field(..., description="Directory containing report templates")
    output_dir: Path = Field(..., description="Directory for generated reports")
    stylesheet_path: Optional[Path] = Field(None, description="Custom CSS for PDF generation")
    
    class Config:
        arbitrary_types_allowed = True

@dataclass
class Template:
    """Report template definition"""
    id: str
    content: str
    sections: List[Dict[str, Any]]
    metadata: Dict[str, Any]

@dataclass
class ReportContent:
    """Generated report content structure"""
    sections: Dict[str, str]
    metadata: Dict[str, Any]

@dataclass
class Report:
    """Complete report with content and metadata"""
    id: UUID
    content: ReportContent
    template_id: str
    created_at: datetime
    version: str
    assessment_id: UUID
    analysis_id: UUID

class TemplateManager:
    """Manages loading and caching of report templates"""
    
    def __init__(self, template_dir: Path):
        self.template_dir = template_dir
        self.template_cache: Dict[str, Template] = {}
    
    async def get_template(self, template_id: str) -> Template:
        """Load and cache template by ID"""
        if template_id not in self.template_cache:
            template_path = self.template_dir / f"{template_id}.jinja2"
            if not template_path.exists():
                raise ValueError(f"Template not found: {template_id}")
            
            # Load template and metadata
            template_content = template_path.read_text()
            metadata_path = template_path.with_suffix('.json')
            metadata = {}
            if metadata_path.exists():
                import json
                metadata = json.loads(metadata_path.read_text())
            
            self.template_cache[template_id] = Template(
                id=template_id,
                content=template_content,
                sections=metadata.get('sections', []),
                metadata=metadata
            )
        
        return self.template_cache[template_id]

class ContentGenerator:
    """Generates report content based on assessment and analysis data"""
    
    async def generate_section(
        self,
        section_config: Dict[str, Any],
        assessment_data: Dict[str, Any],
        analysis_results: Dict[str, Any]
    ) -> str:
        """Generate content for a single report section"""
        section_type = section_config['type']
        
        if section_type == 'summary':
            return self._generate_summary(assessment_data, analysis_results)
        elif section_type == 'assessment':
            return self._generate_assessment_section(assessment_data)
        elif section_type == 'analysis':
            return self._generate_analysis_section(analysis_results)
        elif section_type == 'recommendations':
            return self._generate_recommendations(analysis_results)
        else:
            raise ValueError(f"Unknown section type: {section_type}")
    
    def _generate_summary(self, assessment_data: Dict[str, Any], analysis_results: Dict[str, Any]) -> str:
        # Implement summary generation logic
        pass
    
    def _generate_assessment_section(self, assessment_data: Dict[str, Any]) -> str:
        # Implement assessment section generation logic
        pass
    
    def _generate_analysis_section(self, analysis_results: Dict[str, Any]) -> str:
        # Implement analysis section generation logic
        pass
    
    def _generate_recommendations(self, analysis_results: Dict[str, Any]) -> str:
        # Implement recommendations generation logic
        pass

class PDFGenerator:
    """Handles conversion of report content to PDF format"""
    
    def __init__(self, stylesheet_path: Optional[Path] = None):
        self.stylesheet_path = stylesheet_path
    
    async def generate_pdf(self, report: Report) -> bytes:
        """Convert report content to PDF"""
        try:
            from weasyprint import HTML, CSS
            
            # Convert report content to HTML
            html_content = self._report_to_html(report)
            
            # Apply custom stylesheet if provided
            stylesheets = []
            if self.stylesheet_path and self.stylesheet_path.exists():
                stylesheets.append(CSS(filename=str(self.stylesheet_path)))
            
            # Generate PDF
            html = HTML(string=html_content)
            return html.write_pdf(stylesheets=stylesheets)
        
        except ImportError:
            raise RuntimeError("WeasyPrint is required for PDF generation")
    
    def _report_to_html(self, report: Report) -> str:
        # Implement HTML conversion logic
        pass

class DocumentationAgent(BaseAgent):
    """Agent responsible for generating standardized OT reports"""
    
    def __init__(self, config: DocumentationAgentConfig):
        super().__init__(
            agent_type=AgentType.DOCUMENTATION,
            name="documentation_agent",
            config=config
        )
        self.template_manager = TemplateManager(config.template_dir)
        self.content_generator = ContentGenerator()
        self.pdf_generator = PDFGenerator(config.stylesheet_path)
    
    async def process_assessment(
        self,
        context: AgentContext,
        assessment_data: Dict[str, Any],
        analysis_results: Dict[str, Any],
        template_id: str
    ) -> Report:
        """Generate a report from assessment and analysis data"""
        try:
            # Start processing session
            await self.start_session(context)
            self.update_status(AgentStatus.BUSY)
            
            # Load template
            template = await self.template_manager.get_template(template_id)
            
            # Generate report content
            sections = {}
            for section in template.sections:
                section_content = await self.content_generator.generate_section(
                    section,
                    assessment_data,
                    analysis_results
                )
                sections[section['id']] = section_content
            
            # Create report
            report = Report(
                id=UUID(),
                content=ReportContent(
                    sections=sections,
                    metadata={
                        'template_id': template_id,
                        'generated_at': datetime.utcnow().isoformat(),
                        'context': {
                            'session_id': str(context.session_id),
                            'user_id': str(context.user_id) if context.user_id else None,
                            'therapist_id': str(context.therapist_id) if context.therapist_id else None,
                            'client_id': str(context.client_id) if context.client_id else None
                        }
                    }
                ),
                template_id=template_id,
                created_at=datetime.utcnow(),
                version="1.0",
                assessment_id=UUID(assessment_data['id']),
                analysis_id=UUID(analysis_results['id'])
            )
            
            # Generate PDF
            pdf_content = await self.pdf_generator.generate_pdf(report)
            
            # Save report and PDF
            await self._save_report(report, pdf_content)
            
            return report
            
        except Exception as e:
            await self.handle_error(e, context)
            raise
        
        finally:
            self.update_status(AgentStatus.IDLE)
            await self.end_session(context.session_id)
    
    async def _save_report(self, report: Report, pdf_content: bytes) -> None:
        """Save report content and PDF to storage"""
        # Implement storage logic (database + file system)
        pass