requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
python-dotenv>=1.0.0
alembic>=1.12.1
pydantic>=2.5.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2
asgi-lifespan>=2.1.0
pytest-mock>=3.12.0
freezegun>=1.2.2
factory-boy>=3.3.0
coverage>=7.3.2